Split the Logger out of js-compute-builtins.cpp
